+++ ChangeLog.md
@@ -49,6 +49,7 @@
   - -) 移除遗弃标注项：`FormatStr`、`CliCmd.HasSetting`
   - optimize) `bin.FormatTable` 使用**reflect**扩展器支持任意2维度数组格式化。
   - optimze) `CLI.router` 路由重写将过程化的代码转换为多函数调用，未定义方法注册检测更广泛。
+  - optimize) `CliCmd.parseArgs` 简化代码，使用函数消除冗余，以及使其更好的支持 option 参数的解析。
   - change) 重命名 `CLI.RegisterUnfind` --> `CLI.RegisterAny`
   - **parser**
     - +) `DataReceiver` 新增数据接收器，实现对 json/url等数据格式的解析支持。